+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -13,6 +13,8 @@ group: Deployment & Configs
 
 ### RC
 
+- Added 'Dimension Text' component.
+
 
 {% endcapture %}